json-to-odt/requirements.txt
argparse
PyYAML
yamllint
odfpy
OOoPy
termcolor
colorama
